+++ lib/rspec/activerecord/expectations/query_inspector.rb
@@ -1,7 +1,8 @@
 module RSpec::ActiveRecord::Expectations
   class QueryInspector
     def self.valid_query_types
-      [:queries, :schema_queries, :transaction_queries, :insert_queries]
+      [:queries, :schema_queries, :transaction_queries, :insert_queries,
+       :load_queries, :destroy_queries, :exists_queries]
     end
 
     def categorize(query)
@@ -11,6 +12,14 @@ module RSpec::ActiveRecord::Expectations
         [:transaction_queries]
       elsif query[:name] =~ /Create$/
         [:queries, :insert_queries]
+      elsif query[:name] =~ /Load$/
+        [:queries, :load_queries]
+      elsif query[:name] =~ /Destroy$/
+        [:queries, :destroy_queries]
+      elsif query[:name] =~ /Delete All$/
+        [:queries, :destroy_queries]
+      elsif query[:name] =~ /Exists\?$/
+        [:queries, :exists_queries]
       else
         [:queries]
       end