+++ spec/matchers/query_count_matchers_spec.rb
@@ -23,6 +23,15 @@ RSpec.describe RSpec::ActiveRecord::Expectations::Matchers::QueryCountMatcher do
       expect(matcher.failure_message).to eq("expected block to execute at most 3 queries, but it executed 0")
       expect(matcher.failure_message_when_negated).to eq("expected block not to execute any less than 3 queries, but it executed 0")
     end
+
+    it "has query-type specific output too" do
+      matcher = described_class.new.less_than_or_equal_to(3).insert_queries
+
+      matcher.matches?(example)
+
+      expect(matcher.failure_message).to eq("expected block to execute at most 3 insert queries, but it executed 0")
+      expect(matcher.failure_message_when_negated).to eq("expected block not to execute any less than 3 insert queries, but it executed 0")
+    end
   end
 
   xdescribe "#greater_than"